Consolidate textBlockEncoding tests into the pre-existing file
An earlier slice added a new test file at
src/utils/__tests__/textBlockEncoding.test.ts without noticing the
pre-existing src/utils/textBlockEncoding.test.ts already covered most
of the same ground. Both files were being picked up by vitest,
producing duplicate assertions against the same module.

This commit:
- deletes the duplicate in __tests__/
- folds the uniquely-valuable cases from the duplicate into the
  canonical file:
    * zero-length span (caret-at-position reference)
    * sorts unordered tokens before collapsing into ranges
    * skips malformed segments but keeps valid ones in a mixed string
    * textBlockFromMultipageJson — extracts token indices per page
    * textBlockFromMultipageJson — omits pages with no tokensJsons

Canonical file now 32 cases, all green. Net regression-net impact:
unchanged, just de-duplicated.
